+++ examples/alias/main.tf
@@ -5,7 +5,6 @@ provider "aws" {
   skip_metadata_api_check     = true
   skip_region_validation      = true
   skip_credentials_validation = true
-  skip_requesting_account_id  = true
 }
 
 data "aws_organizations_organization" "this" {}
@@ -74,8 +73,9 @@ module "alias_no_refresh" {
 
   event_source_mapping = {
     sqs = {
-      service          = "sqs"
-      event_source_arn = module.sqs_events.sqs_queue_arn
+      service             = "sqs"
+      event_source_arn    = module.sqs_events.sqs_queue_arn
+      maximum_concurrency = 10
     }
   }