+++ examples/complete/main.tf
@@ -5,7 +5,6 @@ provider "aws" {
   skip_metadata_api_check     = true
   skip_region_validation      = true
   skip_credentials_validation = true
-  skip_requesting_account_id  = true
 }
 
 data "aws_caller_identity" "current" {}
@@ -164,6 +163,16 @@ module "lambda_function" {
       actions   = ["s3:HeadObject", "s3:GetObject"],
       resources = ["arn:aws:s3:::my-bucket/*"]
     }
+  }
+
+  timeouts = {
+    create = "20m"
+    update = "20m"
+    delete = "20m"
+  }
+
+  function_tags = {
+    Language = "python"
   }
 
   tags = {
@@ -361,6 +370,23 @@ module "lambda_function_with_package_deploying_externally" {
   local_existing_package = "../fixtures/python3.8-zip/existing_package.zip"
 
   ignore_source_code_hash = true
+}
+
+####################################################
+# Lambda Function no create log group permission
+####################################################
+
+module "lambda_function_no_create_log_group_permission" {
+  source = "../../"
+
+  function_name = "${random_pet.this.id}-lambda-no-create-log-group-permission"
+  handler       = "index.lambda_handler"
+  runtime       = "python3.8"
+
+  create_package         = false
+  local_existing_package = "../fixtures/python3.8-zip/existing_package.zip"
+
+  attach_create_log_group_permission = false
 }
 
 ###########